make channels into own pallet and tightly couple to xtransfer
