+++ pallets/channels/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+authors = ["PureStake"]
+edition = "2018"
+name = "channels"
+version = "0.1.0"
+description = "manage channels between other parachains"
+
+[dependencies]
+parity-scale-codec = { version = "2.0.0", default-features = false, features = ["derive"] }
+frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+xcm = { git = "https://github.com/paritytech/polkadot", branch = "master", default-features = false }
+cumulus-primitives = { git = "https://github.com/paritytech/cumulus", branch = "master", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+    "parity-scale-codec/std",
+    "frame-support/std",
+    "frame-system/std",
+    "cumulus-primitives/std",
+    "sp-runtime/std",
+    "sp-std/std",
+]
