+++ pallets/parachain-staking/src/lib.rs
@@ -60,7 +60,7 @@ pub use pallet::*;
 
 #[pallet]
 pub mod pallet {
-	use super::{InflationInfo, Range};
+	use crate::inflation::{InflationInfo, Range};
 	use crate::set::OrderedSet;
 	use frame_support::pallet_prelude::*;
 	use frame_support::traits::{Currency, Get, Imbalance, ReservableCurrency};
