+++ pallets/xtransfer/src/lib.rs
@@ -15,23 +15,6 @@
 // along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
 //! Xtransfer handles cross-chain transfers.
-//!
-//! ## Dispatchables
-//! 1. `transfer_to_relay_chain` transfers relay chain tokens to an account on the relay chain
-//! from our parachain
-//! 2. `transfer_to_parachain` transfers tokens to an account on a parachain
-//! Both transfers are sent from accounts on our chain (H160 account id).
-//!
-//! ## Transfers From Other Chains to Self
-//! For transfers from other chains to our chain, their runtime must send us cross chain messages.
-//! 1. For transfers from the relay chain, they queue DownwardMessages. These are processed by the
-//! `DownwardMessageHandlers` associated type for our `cumulus_parachain_system` runtime impl.
-//! 2. For transfers from other parachains, they send us HrmpMessages. These are processed by the
-//! `HrmpMessageHandlers` associated type for our `cumulus_parachain_system` runtime impl.
-//! Both `DownwardMessageHandlers` and `HrmpMessageHandlers` are set to `XcmHandler` in our runtime,
-//! so they use the `MultiCurrency` impl of `TransactAsset` in `./support.rs` to process the
-//! given message. The `Executor` associated type for this pallet uses the same logic to process
-//! the cross-chain messages from our chain.
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
@@ -43,20 +26,13 @@ pub use pallet::*;
 
 #[pallet]
 pub mod pallet {
-	use cumulus_primitives::{
-		relay_chain::Balance as RelayChainBalance, DownwardMessageHandler, InboundDownwardMessage,
-		ParaId,
-	};
+	use cumulus_primitives::{relay_chain::Balance as RelayChainBalance, ParaId};
 	use frame_support::{pallet_prelude::*, traits::Get, transactional};
 	use frame_system::pallet_prelude::*;
 	use sp_runtime::traits::{AtLeast32BitUnsigned, Convert};
-	use sp_std::{convert::TryFrom, prelude::*};
-	use xcm::{
-		v0::{
-			Error as XcmError, ExecuteXcm, Junction, MultiAsset, MultiLocation, NetworkId, Order,
-			OriginKind, SendXcm, Xcm,
-		},
-		VersionedXcm,
+	use sp_std::prelude::*;
+	use xcm::v0::{
+		Error as XcmError, ExecuteXcm, Junction, MultiAsset, MultiLocation, NetworkId, Order, Xcm,
 	};
 	use xcm_executor::traits::LocationConversion;
 
@@ -94,7 +70,7 @@ pub mod pallet {
 
 	/// Configuration trait of this pallet.
 	#[pallet::config]
-	pub trait Config: frame_system::Config {
+	pub trait Config: frame_system::Config + channels::Config {
 		/// Overarching event type
 		type Event: From<Event<Self>> + IsType<<Self as frame_system::Config>::Event>;
 		/// Balances type
@@ -113,41 +89,13 @@ pub mod pallet {
 		type ToMultiLocation: LocationConversion<Self::AccountId>;
 		/// Relay chain identifier
 		type RelayChainNetworkId: Get<NetworkId>;
-		/// Moonbeam parachain identifier
-		type ParaId: Get<ParaId>;
 		/// XCM Executor for executing XCM effects locally
 		type XcmExecutor: ExecuteXcm;
-		/// XCM Sender for sending outgoing messages
-		type XcmSender: SendXcm;
 	}
 
 	#[pallet::event]
 	#[pallet::generate_deposit(fn deposit_event)]
 	pub enum Event<T: Config> {
-		/// Sent channel open request to parachain \[recipient_para_id\]
-		SenderChannelRequested(ParaId),
-		/// Sender channel request accepted by parachain \[recipient_para_id\]
-		SenderChannelAccepted(ParaId),
-		/// Error with received sender channel accepted because one already exists locally
-		SenderChannelAlreadyExists(ParaId),
-		/// Error with recipient channel requested because request already exists locally
-		RecipientChannelAlreadyExists(ParaId),
-		/// Error closing sender channel because channel DNE
-		CloseSenderChannelDNE(ParaId),
-		/// Error closing recipient channel because channel DNE
-		CloseRecipientChannelDNE(ParaId),
-		/// Received new channel request with self as recipient
-		ReceivedRecipientChannelRequest(ParaId),
-		/// Accepted channel open request from parachain
-		AcceptedChannelRequest(ParaId),
-		/// Requested to close the channel with self as sender \[recipient_para_id\]
-		RequestedCloseSenderChannel(ParaId),
-		/// Requested to close the channel with self as recipient \[sender_para_id\]
-		RequestedCloseRecipientChannel(ParaId),
-		/// Closed channel with parachain as recipient and self as sender
-		ClosedSenderChannel(ParaId),
-		/// Closed channel with parachain as sender and self as recipient
-		ClosedRecipientChannel(ParaId),
 		/// Transferred to relay chain. \[src, dest, amount\]
 		TransferredToRelayChain(T::AccountId, AccountId32, T::Balance),
 		/// Transfer to relay chain failed. \[src, dest, amount, error\]
@@ -212,163 +160,11 @@ pub mod pallet {
 		NoRecipientChannelOpen,
 	}
 
-	#[pallet::storage]
-	#[pallet::getter(fn recipient_channel_requests)]
-	/// Open channel requests on the relay chain to self (recipient) from these parachains (sender)
-	pub type RecipientChannelRequests<T: Config> = StorageValue<_, Vec<ParaId>, ValueQuery>;
-
-	#[pallet::storage]
-	#[pallet::getter(fn sender_channels)]
-	/// Stores all para IDs with which this parachain has opened a channel with self as sender
-	pub type SenderChannels<T: Config> = StorageValue<_, Vec<ParaId>, ValueQuery>;
-
-	#[pallet::storage]
-	#[pallet::getter(fn recipient_channels)]
-	/// Stores all para IDs with which this parachain has accepted a channel with self as recipient
-	pub type RecipientChannels<T: Config> = StorageValue<_, Vec<ParaId>, ValueQuery>;
-
 	#[pallet::hooks]
 	impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {}
 
 	#[pallet::call]
 	impl<T: Config> Pallet<T> {
-		#[pallet::weight(10)]
-		/// Request to open HRMP channel with another parachain
-		pub fn open_channel(
-			origin: OriginFor<T>,
-			recipient: ParaId,
-			// temporary until proper Xcm message variant is added with handler logic
-			call: Vec<u8>,
-		) -> DispatchResultWithPostInfo {
-			frame_system::ensure_root(origin)?;
-			let sender = T::ParaId::get();
-			ensure!(sender != recipient, Error::<T>::CannotSendToSelf);
-			// TODO: could check if sender has already made the request; SenderChannelRequests vec
-			let channels = <SenderChannels<T>>::get();
-			ensure!(
-				channels.binary_search(&recipient).is_err(),
-				Error::<T>::MaxOneChannelPerRelation
-			);
-			// call `hrmp_init_open_channel` on relay chain
-			let message = Xcm::Transact {
-				origin_type: OriginKind::Native,
-				call,
-			};
-			// send message to relay chain
-			T::XcmSender::send_xcm(
-				MultiLocation::X1(Junction::Parachain { id: sender.into() }),
-				message,
-			)
-			.map_err(|_| Error::<T>::FailedToSendXcm)?;
-			// emit event
-			Self::deposit_event(Event::SenderChannelRequested(recipient));
-			Ok(().into())
-		}
-		#[pallet::weight(10)]
-		/// Accept a request to open HRMP channel
-		pub fn accept_channel(
-			origin: OriginFor<T>,
-			sender: ParaId,
-			// temporary until proper Xcm message variant is added with handler logic
-			call: Vec<u8>,
-		) -> DispatchResultWithPostInfo {
-			frame_system::ensure_root(origin)?;
-			let self_id = T::ParaId::get();
-			ensure!(sender != self_id, Error::<T>::CannotSendToSelf);
-			// first check if the request even exists (all requests are stored locally)
-			let mut requests = <RecipientChannelRequests<T>>::get();
-			let loc = requests
-				.binary_search(&sender)
-				.map_err(|_| Error::<T>::RecipientRequestDNE)?;
-			requests.remove(loc);
-			let mut channels = <RecipientChannels<T>>::get();
-			ensure!(
-				channels.binary_search(&sender).is_err(),
-				Error::<T>::MaxOneChannelPerRelation
-			);
-			// call `hrmp_accept_open_channel` on relay chain
-			let message = Xcm::Transact {
-				origin_type: OriginKind::Native,
-				call,
-			};
-			// send message to relay chain
-			T::XcmSender::send_xcm(
-				MultiLocation::X1(Junction::Parachain { id: self_id.into() }),
-				message,
-			)
-			.map_err(|_| Error::<T>::FailedToSendXcm)?;
-			// ensured by check further above
-			if let Err(loc) = channels.binary_search(&sender) {
-				channels.insert(loc, sender);
-			}
-			// update recipient channel requests
-			<RecipientChannelRequests<T>>::put(requests);
-			// update recipient channels storage item
-			<RecipientChannels<T>>::put(channels);
-			// emit event
-			Self::deposit_event(Event::AcceptedChannelRequest(sender));
-			Ok(().into())
-		}
-		#[pallet::weight(10)]
-		/// Close an open HRMP channel with self as sender
-		pub fn close_sender_channel(
-			origin: OriginFor<T>,
-			recipient: ParaId,
-			// temporary until hrmp variants added to Xcm
-			call: Vec<u8>,
-		) -> DispatchResultWithPostInfo {
-			frame_system::ensure_root(origin)?;
-			let self_id = T::ParaId::get();
-			ensure!(recipient != self_id, Error::<T>::CannotSendToSelf);
-			let channels = <SenderChannels<T>>::get();
-			ensure!(
-				channels.binary_search(&recipient).is_ok(),
-				Error::<T>::NoSenderChannelOpen
-			);
-			// call `hrmp_close_channel` on relay chain
-			let message = Xcm::Transact {
-				origin_type: OriginKind::Native,
-				call,
-			};
-			// send message to relay chain
-			T::XcmSender::send_xcm(
-				MultiLocation::X1(Junction::Parachain { id: self_id.into() }),
-				message,
-			)
-			.map_err(|_| Error::<T>::FailedToSendXcm)?;
-			Self::deposit_event(Event::RequestedCloseSenderChannel(recipient));
-			Ok(().into())
-		}
-		#[pallet::weight(10)]
-		/// Close an open HRMP channel with self as recipient
-		pub fn close_recipient_channel(
-			origin: OriginFor<T>,
-			sender: ParaId,
-			// temporary until hrmp variants added to Xcm
-			call: Vec<u8>,
-		) -> DispatchResultWithPostInfo {
-			frame_system::ensure_root(origin)?;
-			let self_id = T::ParaId::get();
-			ensure!(sender != self_id, Error::<T>::CannotSendToSelf);
-			let channels = <RecipientChannels<T>>::get();
-			ensure!(
-				channels.binary_search(&sender).is_ok(),
-				Error::<T>::NoRecipientChannelOpen
-			);
-			// call is to `hrmp_close_channel` on relay chain
-			let message = Xcm::Transact {
-				origin_type: OriginKind::Native,
-				call,
-			};
-			// send message to accept the channel request
-			T::XcmSender::send_xcm(
-				MultiLocation::X1(Junction::Parachain { id: self_id.into() }),
-				message,
-			)
-			.map_err(|_| Error::<T>::FailedToSendXcm)?;
-			Self::deposit_event(Event::RequestedCloseRecipientChannel(sender));
-			Ok(().into())
-		}
 		/// Transfer relay chain tokens to relay chain.
 		#[pallet::weight(10)]
 		#[transactional]
@@ -426,7 +222,7 @@ pub mod pallet {
 			let who = ensure_signed(origin)?;
 			ensure!(para_id != T::ParaId::get(), Error::<T>::CannotSendToSelf);
 			ensure!(
-				<SenderChannels<T>>::get().binary_search(&para_id).is_ok(),
+				<channels::Pallet<T>>::sender_channels().contains(&para_id),
 				Error::<T>::NoSenderChannelOpen
 			);
 
@@ -496,7 +292,7 @@ pub mod pallet {
 			let who = ensure_signed(origin)?;
 			ensure!(para_id != T::ParaId::get(), Error::<T>::CannotSendToSelf);
 			ensure!(
-				<SenderChannels<T>>::get().binary_search(&para_id).is_ok(),
+				<channels::Pallet<T>>::sender_channels().contains(&para_id),
 				Error::<T>::NoSenderChannelOpen
 			);
 
@@ -550,77 +346,6 @@ pub mod pallet {
 			}
 
 			Ok(().into())
-		}
-	}
-
-	impl<T: Config> DownwardMessageHandler for Module<T> {
-		fn handle_downward_message(msg: InboundDownwardMessage) {
-			match VersionedXcm::decode(&mut &msg.msg[..]).map(Xcm::try_from) {
-				Ok(Ok(xcm)) => {
-					match xcm {
-						Xcm::HrmpNewChannelOpenRequest { sender, .. } => {
-							let mut channels = <RecipientChannelRequests<T>>::get();
-							let sender: ParaId = sender.into();
-							// if request from id already exists, not added to requests
-							if let Err(loc) = channels.binary_search(&sender) {
-								channels.insert(loc, sender);
-								<RecipientChannelRequests<T>>::put(channels);
-								Self::deposit_event(Event::ReceivedRecipientChannelRequest(sender));
-							} else {
-								// event error
-								Self::deposit_event(Event::RecipientChannelAlreadyExists(sender));
-							}
-						}
-						Xcm::HrmpChannelAccepted { recipient } => {
-							let mut channels = <SenderChannels<T>>::get();
-							let recipient: ParaId = recipient.into();
-							// if channel with id already exists, not added to channels
-							if let Err(loc) = channels.binary_search(&recipient) {
-								channels.insert(loc, recipient);
-								<SenderChannels<T>>::put(channels);
-								Self::deposit_event(Event::SenderChannelAccepted(recipient));
-							} else {
-								// event error
-								Self::deposit_event(Event::SenderChannelAlreadyExists(recipient));
-							}
-						}
-						Xcm::HrmpChannelClosing {
-							sender, recipient, ..
-						} => {
-							let self_id = T::ParaId::get();
-							let sender: ParaId = sender.into();
-							let recipient: ParaId = recipient.into();
-							if sender == self_id {
-								let mut channels = <SenderChannels<T>>::get();
-								// update storage
-								if let Ok(loc) = channels.binary_search(&recipient) {
-									channels.remove(loc);
-									<SenderChannels<T>>::put(channels);
-									Self::deposit_event(Event::ClosedSenderChannel(recipient));
-								} else {
-									// event error
-									Self::deposit_event(Event::CloseSenderChannelDNE(recipient));
-								}
-							}
-							if recipient == self_id {
-								let mut channels = <RecipientChannels<T>>::get();
-								// update storage
-								if let Ok(loc) = channels.binary_search(&sender) {
-									channels.remove(loc);
-									<RecipientChannels<T>>::put(channels);
-									Self::deposit_event(Event::ClosedRecipientChannel(sender));
-								} else {
-									// event error
-									Self::deposit_event(Event::CloseRecipientChannelDNE(recipient));
-								}
-							}
-						}
-						_ => (),
-					}
-				}
-				Ok(Err(..)) => (),
-				Err(..) => (),
-			}
 		}
 	}
